backend/requirements.txt
coreapi==2.1.1
dj-database-url==0.4.1
Django==1.10.4
django-rest-swagger==2.1.0
djangorestframework==3.5.3
gunicorn==19.6.0
itypes==1.1.0
openapi-codec==1.2.1
psycopg2==2.5.4
pytz==2016.10
requests==2.12.4
simplejson==3.10.0
uritemplate==3.0.0
whitenoise==3.2.2
